Fixed empty Deck issue
Before, the program would check for an empty Deck and then 2 AI's would
play a card, so if the Deck had one card in it, the 2nd AI player would
attempt to use a card from an empty Deck.
Placed the empty Deck check before each AITurn call and before the
Player's turn starts again.